requirements.txt
yt-dlp
scenedetect[opencv]
torch
torchvision
transformers==4.41.2
einops
rapidfuzz
prompt_toolkit
pillow
google-generativeai
accelerate
python-dotenv